complete order submission flow for logged in users
