+++ pages/account.tsx
@@ -4,7 +4,6 @@ import { useState } from 'react'
 import { useEffect } from 'react'
 import { Layout } from '../components/General'
 import { EmptyStatePlaceholder } from '../components/Shop'
-import { useCartContext } from '../context/CartContext'
 import { CartItem } from '../types'
 import { formatDate, addToDate } from '../helpers/utilities'
 import Link from 'next/link'
@@ -25,421 +24,7 @@ const Account = () => {
   const { data: session, status } = useSession()
   const [csrfToken, setCsrfToken] = useState('')
 
-  const cart = useCartContext()
-
-  const testOrders: Order[] = [
-    {
-      orderId: '12234783',
-      orderDate: new Date('2021-09-01'),
-      orderDetails: {
-        subtotal: 100,
-        shipping: 10,
-        tax: 0,
-        total: 110
-      },
-      items: [
-        {
-          product: {
-            coffeeBagSizes: [
-              {
-                bagSize: {
-                  id: 12,
-                  name: 'M',
-                  volumeDiscount: 0.1,
-                  weightInGrams: 500
-                },
-                quantity: 6
-              },
-              {
-                bagSize: {
-                  id: 14,
-                  name: 'XL',
-                  volumeDiscount: 0.2,
-                  weightInGrams: 1000
-                },
-                quantity: 1
-              },
-              {
-                bagSize: {
-                  id: 13,
-                  name: 'L',
-                  volumeDiscount: 0.15,
-                  weightInGrams: 750
-                },
-                quantity: 5
-              },
-              {
-                bagSize: {
-                  id: 11,
-                  name: 'S',
-                  volumeDiscount: 0,
-                  weightInGrams: 250
-                },
-                quantity: 1
-              }
-            ],
-            description:
-              "This signature House Blend was crafted as the hallmark coffee for Irving's original cafe at 71 Irving Place. Classic, rich, smooth flavors are great with milk.",
-            flavor: 'Sweet & Smooth',
-            flavorNotes: [
-              {
-                flavorNote: 'Spice',
-                id: 8
-              },
-              {
-                flavorNote: 'Nut',
-                id: 10
-              },
-              {
-                flavorNote: 'Citrus',
-                id: 2
-              }
-            ],
-            id: 35,
-            imageUrl:
-              'https://res.cloudinary.com/roastcollective/image/upload/h_1000,w_1000,f_auto,fl_progressive:steep,q_auto:good/v1661179176/solidus/nien9l92dj3bis3vbvof.png',
-            location: 'New York, NY',
-            name: '71 House Blend',
-            pricePerKilo: 19.65,
-            roastLevel: 6,
-            roaster: 'Irving Farm',
-            roasterNotes:
-              'Our signature House Blend was created as the hallmark coffee for our original cafe at 71 Irving Place. The blend has a chocolate, round, and smooth flavor.'
-          },
-          quantity: 1,
-          size: {
-            bagSize: {
-              id: 13,
-              name: 'L',
-              volumeDiscount: 0.15,
-              weightInGrams: 750
-            },
-            quantity: 5
-          }
-        },
-        {
-          product: {
-            coffeeBagSizes: [
-              {
-                bagSize: {
-                  id: 13,
-                  name: 'L',
-                  volumeDiscount: 0.15,
-                  weightInGrams: 750
-                },
-                quantity: 5
-              },
-              {
-                bagSize: {
-                  id: 12,
-                  name: 'M',
-                  volumeDiscount: 0.1,
-                  weightInGrams: 500
-                },
-                quantity: 1
-              },
-              {
-                bagSize: {
-                  id: 14,
-                  name: 'XL',
-                  volumeDiscount: 0.2,
-                  weightInGrams: 1000
-                },
-                quantity: 2
-              },
-              {
-                bagSize: {
-                  id: 11,
-                  name: 'S',
-                  volumeDiscount: 0,
-                  weightInGrams: 250
-                },
-                quantity: 1
-              }
-            ],
-            description:
-              'Deeply sweet and chocolaty, with a kiss of roastiness pairing beautifully with notes of toasted walnut.',
-            flavor: 'Comforting & Rich',
-            flavorNotes: [
-              {
-                flavorNote: 'Spice',
-                id: 8
-              },
-              {
-                flavorNote: 'Ripe Fruit',
-                id: 3
-              },
-              {
-                flavorNote: 'Brow Sugar',
-                id: 4
-              },
-              {
-                flavorNote: 'Milk Chocolate',
-                id: 1
-              }
-            ],
-            id: 32,
-            imageUrl:
-              'https://res.cloudinary.com/roastcollective/image/upload/h_1000,w_1000,f_auto,fl_progressive:steep,q_auto:good/v1660237550/solidus/fdaeef3wch3cnmzqommf.png',
-            location: 'Bentonville, AR',
-            name: 'Black Bear Blend',
-            pricePerKilo: 25.45,
-            roastLevel: 7,
-            roaster: 'Airship',
-            roasterNotes:
-              "A lot can be (and has been) said about dark roasted coffee. This is the only coffee we profile by roast color. But we don't just turn coffee dark by roasting it longer. We pay close attention to Black Bear by selecting the best coffee and roasting it carefully so it retains its sweetness. Black Bear is complex, velvety, and roasted darker to bring out a rich body."
-          },
-          quantity: 2,
-          size: {
-            bagSize: {
-              id: 14,
-              name: 'XL',
-              volumeDiscount: 0.2,
-              weightInGrams: 1000
-            },
-            quantity: 6
-          }
-        },
-        {
-          product: {
-            coffeeBagSizes: [
-              {
-                bagSize: {
-                  id: 13,
-                  name: 'L',
-                  volumeDiscount: 0.15,
-                  weightInGrams: 750
-                },
-                quantity: 3
-              },
-              {
-                bagSize: {
-                  id: 11,
-                  name: 'S',
-                  volumeDiscount: 0,
-                  weightInGrams: 250
-                },
-                quantity: 6
-              },
-              {
-                bagSize: {
-                  id: 14,
-                  name: 'XL',
-                  volumeDiscount: 0.2,
-                  weightInGrams: 1000
-                },
-                quantity: 2
-              },
-              {
-                bagSize: {
-                  id: 12,
-                  name: 'M',
-                  volumeDiscount: 0.1,
-                  weightInGrams: 500
-                },
-                quantity: 0
-              }
-            ],
-            description:
-              'With a big, chocolaty body, cozy roasted almond nuttiness and tons of balanced caramel sweetness, this Trade-exclusive cup is just about as comforting as you can get.',
-            flavor: 'Comforting & Rich',
-            flavorNotes: [
-              {
-                flavorNote: 'Ripe Fruit',
-                id: 3
-              },
-              {
-                flavorNote: 'Brow Sugar',
-                id: 4
-              },
-              {
-                flavorNote: 'Roastiness',
-                id: 5
-              },
-              {
-                flavorNote: 'Milk Chocolate',
-                id: 1
-              }
-            ],
-            id: 26,
-            imageUrl:
-              'https://res.cloudinary.com/roastcollective/image/upload/h_1000,w_1000,f_auto,fl_progressive:steep,q_auto:good/v1660249589/solidus/rkbqu9yice0f4nuuis7x.png',
-            location: 'Denver, CO',
-            name: 'Bom Senso',
-            pricePerKilo: 18.4,
-            roastLevel: 8,
-            roaster: 'Huckleberry',
-            roasterNotes:
-              'Bom Senso is a Trade-exclusive, traditional coffee. We source straightforward coffees from Latin America, and roast them a bit darker (for Huckleberry, at least) to create a big, bold cup for the dark roast lovers out there.'
-          },
-          quantity: 1,
-          size: {
-            bagSize: {
-              id: 13,
-              name: 'L',
-              volumeDiscount: 0.15,
-              weightInGrams: 750
-            },
-            quantity: 5
-          }
-        },
-        {
-          product: {
-            coffeeBagSizes: [
-              {
-                bagSize: {
-                  id: 12,
-                  name: 'M',
-                  volumeDiscount: 0.1,
-                  weightInGrams: 500
-                },
-                quantity: 3
-              },
-              {
-                bagSize: {
-                  id: 13,
-                  name: 'L',
-                  volumeDiscount: 0.15,
-                  weightInGrams: 750
-                },
-                quantity: 2
-              },
-              {
-                bagSize: {
-                  id: 14,
-                  name: 'XL',
-                  volumeDiscount: 0.2,
-                  weightInGrams: 1000
-                },
-                quantity: 0
-              },
-              {
-                bagSize: {
-                  id: 11,
-                  name: 'S',
-                  volumeDiscount: 0,
-                  weightInGrams: 250
-                },
-                quantity: 7
-              }
-            ],
-            description:
-              "A deep black cherry sweetness peeks through this coffee's pronounced dark chocolate and cinnamon notes.",
-            flavor: 'Sweet & Smooth',
-            flavorNotes: [
-              {
-                flavorNote: 'Roastiness',
-                id: 5
-              }
-            ],
-            id: 25,
-            imageUrl:
-              'https://res.cloudinary.com/roastcollective/image/upload/h_1000,w_1000,f_auto,fl_progressive:steep,q_auto:good/v1663185414/solidus/xw8crsq1xams1qcgds4b.png',
-            location: 'Nashville, TN',
-            name: 'Canopy',
-            pricePerKilo: 25.85,
-            roastLevel: 6,
-            roaster: 'Common Voice',
-            roasterNotes:
-              'Rise above it all with this rich, dark roast with a strong,complex flavor from Latin America. Deliciously aromatic, ahint of Chocolate, a touch of Spice, and notes of Caramelwill lift your spirits when you need to go bold.'
-          },
-          quantity: 2,
-          size: {
-            bagSize: {
-              id: 13,
-              name: 'L',
-              volumeDiscount: 0.15,
-              weightInGrams: 750
-            },
-            quantity: 5
-          }
-        },
-        {
-          product: {
-            coffeeBagSizes: [
-              {
-                bagSize: {
-                  id: 12,
-                  name: 'M',
-                  volumeDiscount: 0.1,
-                  weightInGrams: 500
-                },
-                quantity: 6
-              },
-              {
-                bagSize: {
-                  id: 14,
-                  name: 'XL',
-                  volumeDiscount: 0.2,
-                  weightInGrams: 1000
-                },
-                quantity: 1
-              },
-              {
-                bagSize: {
-                  id: 13,
-                  name: 'L',
-                  volumeDiscount: 0.15,
-                  weightInGrams: 750
-                },
-                quantity: 5
-              },
-              {
-                bagSize: {
-                  id: 11,
-                  name: 'S',
-                  volumeDiscount: 0,
-                  weightInGrams: 250
-                },
-                quantity: 1
-              }
-            ],
-            description:
-              "This signature House Blend was crafted as the hallmark coffee for Irving's original cafe at 71 Irving Place. Classic, rich, smooth flavors are great with milk.",
-            flavor: 'Sweet & Smooth',
-            flavorNotes: [
-              {
-                flavorNote: 'Spice',
-                id: 8
-              },
-              {
-                flavorNote: 'Nut',
-                id: 10
-              },
-              {
-                flavorNote: 'Citrus',
-                id: 2
-              }
-            ],
-            id: 35,
-            imageUrl:
-              'https://res.cloudinary.com/roastcollective/image/upload/h_1000,w_1000,f_auto,fl_progressive:steep,q_auto:good/v1661179176/solidus/nien9l92dj3bis3vbvof.png',
-            location: 'New York, NY',
-            name: '71 House Blend',
-            pricePerKilo: 19.65,
-            roastLevel: 6,
-            roaster: 'Irving Farm',
-            roasterNotes:
-              'Our signature House Blend was created as the hallmark coffee for our original cafe at 71 Irving Place. The blend has a chocolate, round, and smooth flavor.'
-          },
-          quantity: 4,
-          size: {
-            bagSize: {
-              id: 12,
-              name: 'M',
-              volumeDiscount: 0.1,
-              weightInGrams: 500
-            },
-            quantity: 5
-          }
-        }
-      ]
-    }
-  ]
-
-  console.log(cart.cart)
-
-  const [orders, setOrders] = useState(testOrders)
+  const [orders, setOrders] = useState<Order[]>([])
 
   console.log(orders)
 
@@ -453,15 +38,17 @@ const Account = () => {
     getAuth()
 
     if (status === 'authenticated') {
-      fetch('/api/orders', {
+      fetch('/api/checkout-service/orders', {
         method: 'GET',
         headers: {
           'Content-Type': 'application/json',
-          'X-CSRF-Token': csrfToken
+          'X-CSRF-Token': csrfToken,
+          Authorization: 'Bearer ' + session?.accessToken
         }
       })
         .then(response => response.json())
         .then(data => {
+          console.log(data)
           setOrders(data)
         })
     }
@@ -496,7 +83,7 @@ const Account = () => {
             />
           ) : (
             <div className='flex flex-col gap-y-2 pt-4'>
-              {orders.map(order => (
+              {orders?.map(order => (
                 <div key={order.orderId} className='rounded-md border border-zinc-200 p-8'>
                   <div className='flex items-center justify-between'>
                     <div className='flex items-center gap-x-4'>
@@ -509,7 +96,7 @@ const Account = () => {
                   </div>
 
                   <div className='grid grid-cols-3 gap-4 pt-4'>
-                    {order.items.map(item => (
+                    {order?.items?.map(item => (
                       <Link
                         key={item.product.id + item.size.bagSize.id}
                         href={`/products/${item.product.id}`}>
@@ -538,6 +125,16 @@ const Account = () => {
             </div>
           )}
         </section>
+      </Layout>
+    )
+  }
+
+  if (status === 'loading') {
+    return (
+      <Layout>
+        <div className='flex min-h-[85vh] items-center justify-center'>
+          <div className='h-16 w-16 animate-spin rounded-full border-b-2 border-amber-500'></div>
+        </div>
       </Layout>
     )
   }
